Added validation for Uris in MessageRegistryFile
